Working code! Changed how everything works.
This is considered as Alpha but it works pretty great.

For Raspberry PI (Considerated Data is plugged on GPIO 7)
-arduino_wireless
Sends a number to the Arduino.
Ex: ./arduino_wireless 50

-ardwl_cs.sh (arduino_wireless_changestate)
Send digital pin and state  to the arduino with arduino_wireless (see 433wirelessPI.ino)
Ex: ardwl_cs.sh 5 1 (means change digital Pin 5 to state 1)

-The arduino is listening for 433Mhz Message on PIN 2
If it received a message 5 and 1 for example
Then it changes to digitalPin 5 to stated High

There is little verification but nothing serious, you should consider
that things could go wrong.
For example: If something parasitated 433.92, I don't know what will happen.

Next steps:
- Make a library on the Arduino

- Verification of communication (Maybe a number could be used between
each number to verify if everything goes well).

- ID for using multiple Arduino (Would be cool but I would need MOAR
arduino for testing purpose)

- Inverse communication (Arduino speaks to RPI)

Feel free to contribute or ask for functionnality! (or troll).
I would really appreciate to have feedback on this to improve my code.

Anyway it works pretty great. I tested a script that light up leds
and shut them down and the distance was quite long (despite using a
cable as antenna)